+++ macmo/Sources/presentation/search/SearchMemoView.swift
@@ -1,0 +1,18 @@
+//
+//  SearchMemoView.swift
+//  macmo
+//
+//  Created by 신동규 on 9/28/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct SearchMemoView: View {
+    
+    @StateObject var model: SearchMemoViewModel
+    
+    var body: some View {
+        Text("SearchView")
+    }
+}
